refactor(queue-manager): move swap-execution codes out of wallets-shared
The queue manager reached into a wallet package for swap-execution logic. Each
piece now sits with whoever owns it, and neither the queue manager nor its
preset imports wallets-shared any more.

Most of what it pulled was a re-declaration of wallets-core/legacy rather than
original code, so WalletType, Network, WalletState and Providers are repoints to
the types that already existed there. Networks comes from wallets-blockchains,
which also gains HYPERLIQUID_SIGN_NETWORK — Rango routing knowledge that had no
business in a wallet package, and which the widget reads too. Meta and
AllBlockchains move into the preset, the lower of the two packages that use them.

wallets-shared no longer stands between those owners and their consumers: its
Networks re-export and its duplicate `Network = string` alias are gone, so the
widget reads Networks from wallets-blockchains and Network from
wallets-core/legacy exactly as the queue manager already does.

getEvmProvider is deleted rather than moved: every provider registers only the
hub version, so the legacy branch it guarded was unreachable. Its now-unused
argument stays behind a TODO so the context shape is a separate change.
