+++ queue-manager/queue-manager-demo/package.json
@@ -34,9 +34,9 @@
     "@rango-dev/provider-trustwallet": "^0.63.0",
     "@rango-dev/queue-manager-rango-preset": "^0.64.0",
     "@rango-dev/queue-manager-react": "^0.34.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-react": "^0.48.0",
-    "@rango-dev/wallets-shared": "^0.62.0",
     "bignumber.js": "^9.1.1",
     "crypto-browserify": "^3.12.1",
     "ethers": "^6.13.2",
