+++ queue-manager/rango-preset/package.json
@@ -24,7 +24,6 @@
     "@rango-dev/queue-manager-core": "*",
     "@rango-dev/queue-manager-react": "*",
     "@rango-dev/wallets-core": "*",
-    "@rango-dev/wallets-shared": "*",
     "bignumber.js": "*",
     "rango-sdk": "*",
     "rango-types": "*",
@@ -39,6 +38,7 @@
     "@hub3js/stellar": "0.0.0-experimental-50f6f95a-20260819",
     "@hub3js/xrpl": "0.0.0-experimental-50f6f95a-20260819",
     "@rango-dev/logging-core": "^0.14.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@stellar/stellar-sdk": "^15.0.1",
     "rango-types": "^0.5.0",
     "ts-results": "^3.3.0",
