+++ queue-manager/rango-preset/src/types.ts
@@ -6,19 +6,18 @@ import type {
   QueueDef,
   QueueStorage,
 } from '@rango-dev/queue-manager-core';
+import type { EvmNetworksChainInfo } from '@rango-dev/wallets-blockchains';
 import type {
   LegacyConnectResult as ConnectResult,
+  LegacyNetwork as Network,
+  LegacyProviders as Providers,
+  LegacyState as WalletState,
   LegacyWalletType as WalletType,
 } from '@rango-dev/wallets-core/legacy';
-import type {
-  Meta,
-  Network,
-  Providers,
-  WalletState,
-} from '@rango-dev/wallets-shared';
 import type { Transaction } from 'rango-sdk';
 import type {
   APIErrorCode,
+  BlockchainMeta,
   EvmBlockchainMeta,
   PendingSwap,
   PendingSwapStep,
@@ -28,6 +27,15 @@ import type {
 export type RemoveNameField<T, U extends string> = {
   [Property in keyof T as Exclude<Property, U>]: T[Property];
 };
+
+export type AllBlockchains = { [key: string]: BlockchainMeta };
+
+export interface Meta {
+  blockchains: AllBlockchains;
+  evmNetworkChainInfo: EvmNetworksChainInfo;
+  getSupportedChainNames: (type: WalletType) => Network[] | null;
+  evmBasedChains: EvmBlockchainMeta[];
+}
 
 export type ArrayElement<A> = A extends readonly (infer T)[] ? T : never;
