+++ widget/embedded/src/utils/wallets.ts
@@ -5,12 +5,14 @@ import type {
   Wallet,
   WalletInfoWithExtra,
 } from '../types';
-import type { LegacyWalletType as WalletType } from '@rango-dev/wallets-core/legacy';
+import type {
+  LegacyNetwork as Network,
+  LegacyWalletType as WalletType,
+} from '@rango-dev/wallets-core/legacy';
 import type {
   ExtendedWalletInfo,
   ProviderContext,
 } from '@rango-dev/wallets-react';
-import type { Network } from '@rango-dev/wallets-shared';
 import type { BlockchainMeta, Token, TransactionType } from 'rango-sdk';
 
 import {
@@ -19,10 +21,13 @@ import {
   WalletState as WalletStatus,
 } from '@rango-dev/ui';
 import {
+  HYPERLIQUID_SIGN_NETWORK,
+  Networks,
+} from '@rango-dev/wallets-blockchains';
+import {
   legacyGetBlockChainNameFromId as getBlockChainNameFromId,
   legacyReadAccountAddress as readAccountAddress,
 } from '@rango-dev/wallets-core/legacy';
-import { HYPERLIQUID_SIGN_NETWORK, Networks } from '@rango-dev/wallets-shared';
 import BigNumber from 'bignumber.js';
 import { isAddress as isEvmAddress } from 'ethers';
